feat: add auth support
