+++ gcp/Dockerfile
@@ -1,10 +1,11 @@
 # syntax=docker/dockerfile:1
-FROM python:3.9-slim
+FROM gcr.io/google.com/cloudsdktool/cloud-sdk:slim
 
 ENV PYTHONUNBUFFERED True
+ENV GOOGLE_APPLICATION_CREDENTIALS ./creds.json
 
 COPY requirements.txt .
-RUN pip install -r requirements.txt
+RUN pip3 install -r requirements.txt
 
 ENV APP_HOME /app
 WORKDIR ${APP_HOME}
