+++ gcp/pyAnVIL/anvil/terra/api.py
@@ -6,7 +6,7 @@ import logging
 import re
 from attrdict import AttrDict
 
-logger = logging.getLogger('anvil.terra.api')
+logger = logging.getLogger("anvil.terra.api")
 
 USER_PROJECT = None
 
@@ -34,30 +34,44 @@ def get_projects(namespaces=None, project_pattern=None):
 
     """
     workspaces = FAPI.list_workspaces()
-    print(f"WORK: {workspaces}")
     workspaces = workspaces.json()
 
     if namespaces:
-        workspaces = [AttrDict(w) for w in workspaces if w['workspace']['namespace'] in namespaces]
+        workspaces = [
+            AttrDict(w)
+            for w in workspaces
+            if w["workspace"]["namespace"] in namespaces
+        ]
 
     if project_pattern:
-        workspaces = [AttrDict(w) for w in workspaces if re.match(project_pattern, w['workspace']['name'], re.IGNORECASE)]
+        workspaces = [
+            AttrDict(w)
+            for w in workspaces
+            if re.match(project_pattern, w["workspace"]["name"], re.IGNORECASE)
+        ]
 
     # normalize fields
     for w in workspaces:
-        if 'project_files' not in w.workspace:
+        if "project_files" not in w.workspace:
             w.workspace.project_files = []
     return workspaces
 
 
 @memoize
-def get_entities(namespace='anvil-datastorage', workspace=None, entity_name=None):
+def get_entities(
+    namespace="anvil-datastorage", workspace=None, entity_name=None
+):
     """Return all entities in a workspace."""
-    entities = [AttrDict(e) for e in FAPI.get_entities(namespace, workspace, entity_name).json()]
+    entities = [
+        AttrDict(e)
+        for e in FAPI.get_entities(namespace, workspace, entity_name).json()
+    ]
     return entities
 
 
 @memoize
 def get_schema(namespace, workspace):
     """Fetch all entity types."""
-    return FAPI.list_entity_types(namespace=namespace, workspace=workspace).json()
+    return FAPI.list_entity_types(
+        namespace=namespace, workspace=workspace
+    ).json()
